M7.9d: hoist goal-spec types + validator into packages/shared
Closes the last duplication flagged in M3.2 / M7.8 commit messages.
The constants + validator algorithm now live in @espace-devhub/shared/
goal-specs as the single source of truth; both apps re-export from
their feature barrels so existing import sites keep working.

What ships
  packages/shared/src/goal-specs/
    types.js + types.d.ts       — SPEC_KINDS, SOURCE_METRICS, SPEC_VARIANTS,
                                   SOURCE_PROVIDERS, SOURCE_WINDOWS,
                                   MANUAL_CADENCES, TARGET_OPS,
                                   CONTEXT_QUESTION_KINDS, DELEGATED_JUDGES,
                                   SPEC_KIND_META, SPEC_SCHEMA_VERSION,
                                   normalizeCadence + cadence synonym table.
                                   JS runtime + .d.ts sidecar gives TS the
                                   literal types without a build step.
    validator.js + validator.d.ts
                                 — validateSpec, isSpec, buildSpec. Same
                                   strict-on-required, permissive-on-
                                   optional algorithm both sides had.
    index.js + index.d.ts       — public subpath barrel.

  packages/shared/package.json
    "type": "module", subpath exports for "." and "./goal-specs",
    "types" + "default" conditions so TS picks up the .d.ts.

  apps/web/package.json
  apps/api/package.json
    Add "@espace-devhub/shared": "*" workspace dep.

  apps/web/src/features/goal-specs/
    index.js                    — re-exports the shared symbols so the
                                   feature's public surface is unchanged.
    specs-store.js              — imports validateSpec from shared.
    use-goal-specs.js           — same.
    specs-sync.js               — comment refresh.
    types.js                    — DELETED (moved to shared).
    schema.js                   — DELETED (moved to shared as validator.js).

  apps/api/src/modules/ai/classifier/
    events.ts                   — imports ValidatedSpec type from shared.
    mistral-classifier.ts       — imports SPEC_KIND_META, SPEC_VARIANTS,
                                   SPEC_KINDS, validateSpec from shared.
    spec-types.ts               — DELETED.
    spec-validator.ts           — DELETED.

  apps/api/src/modules/goal-specs/controller.ts
  apps/api/src/modules/migrate/controller.ts
    Import validateSpec + ValidatedSpec from shared. Comment refresh.

  apps/api/src/db/types.ts
  apps/api/src/db/schemas/goal-specs.schema.ts
    Comment refresh pointing at the new home.

Drive-by
  .gitignore                    — add **/dist. The api's tsc build emits
                                   to apps/api/dist/; was never tracked
                                   but also wasn't ignored, so the build
                                   verification step in this PR surfaced
                                   it as noise on every contributor's
                                   `git status` after running build:api.

Verified by
  npm install               — workspaces link @espace-devhub/shared
                              into root node_modules; both apps see it.
  npm run typecheck:api    — clean.
  npm run build:web        — clean.
  npm run build:api        — clean. apps/api/dist resolves the package
                              via the symlinked node_modules at runtime.
  Node smoke check         — `import("@espace-devhub/shared/goal-specs")`
                              from apps/api resolves all 19 exports.

What this unblocks
  - Any future addition to spec-kind / cadence / metric vocabulary is
    a one-file edit. The two sides cannot drift again.
  - The shared package now has a real module to anchor the next round
    of hoists (date / week helpers identified in M-later notes).
