+++ apps/api/package.json
@@ -16,6 +16,7 @@
     "admin:create": "tsx scripts/admin-create.ts"
   },
   "dependencies": {
+    "@espace-devhub/shared": "*",
     "@node-rs/argon2": "^2.0.2",
     "cookie-parser": "^1.4.7",
     "cors": "^2.8.5",
