+++ apps/web/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@api/nim": "file:.api/apis/nim",
+    "@espace-devhub/shared": "*",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-dialog": "^1.1.15",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
